pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "flow-grpo"
version = "0.0.1"
description = "Training Flow Matching Models via Online RL"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10,<3.12"
authors = [
    { name = "Flow-GRPO Team" }
]

dependencies = [
    # Core Deep Learning
    "torch==2.6.0",
    "torchvision==0.21.0",
    "torchaudio",
    "transformers>=4.44.0",
    "accelerate==1.4.0",
    "diffusers==0.33.1",
    
    # Data Processing
    "numpy==1.26.4",
    "pandas==2.2.3",
    "scipy==1.15.2",
    "scikit-learn==1.6.1",
    "scikit-image==0.25.2",
    
    # Image Processing
    "albumentations==1.4.10",
    "opencv-python==4.11.0.86",
    "pillow==10.4.0",
    
    # Utilities
    "tqdm==4.67.1",
    "wandb==0.18.7",
    "pydantic==2.10.6",
    "requests==2.32.3",
    "matplotlib==3.10.0",
    
    # Training Optimization
    "deepspeed==0.16.4",
    "peft==0.10.0",
    "bitsandbytes==0.45.3",
    
    # API & Server
    "aiohttp==3.11.13",
    "fastapi==0.115.11",
    "uvicorn==0.34.0",
    
    # HuggingFace
    "huggingface-hub==0.29.1",
    "datasets==3.3.2",
    "tokenizers==0.19.1",
    
    # Miscellaneous
    "einops==0.8.1",
    "nvidia-ml-py==12.570.86",
    "xformers",
    "absl-py",
    "ml_collections",
    "sentencepiece",
    "openai",
]

[project.optional-dependencies]
# Development dependencies
dev = [
    "ipython==8.34.0",
    "black==24.2.0",
    "pytest==8.2.0",
]

# Reward Models - Choose what you need
#geneval = [
#    "geneval>=0.1.0",
#]

ocr = [
    "paddlepaddle-gpu==2.6.2",
    "paddleocr==2.9.1",
    "python-Levenshtein",
]

imagereward = [
    "image-reward",
]

#unified-reward = [
#    "sglang[all]",
#]

# All reward models (for experimentation)
#all-rewards = [
#    "geneval>=0.1.0",
#    "paddlepaddle-gpu==2.6.2",
#    "paddleocr==2.9.1",
#    "python-Levenshtein",
#    "image-reward",
#]

# Optional optimizations
flash-attn = [
    "flash-attn==2.7.4.post1",
]

bagel = [
    "transformers>=4.44.0",
    "flash-attn==2.7.4.post1",
]

[dependency-groups]
dev = [
    "ipython==8.34.0",
    "black==24.2.0",
    "pytest==8.2.0",
]

[tool.uv]
# UV-specific configuration
# Allow to install all optional dependencies at once
default-groups = ["dev"]

[tool.setuptools]
packages = ["flow_grpo"]
